+++ nepali/datetime/__init__.py
@@ -2,6 +2,7 @@ from ._datetime import nepalidate, nepalitime, nepalidatetime
 from ._formatter import NepaliDateTimeFormatter
 from ._humanize import HumanizeDateTime, nepalihumanize
 from ._nepalimonth import nepalimonth
+from ._nepaliweek import nepaliweek
 
 __all__ = [
     "nepalidate",
@@ -9,6 +10,7 @@ __all__ = [
     "nepalidatetime",
     "nepalihumanize",
     "nepalimonth",
+    "nepaliweek",
     "NepaliDateTimeFormatter",
     "HumanizeDateTime",
 ]